Separate the library and binary roots
