+++ mm2src/mm2_lib.rs
@@ -1,0 +1,20 @@
+#![feature(non_ascii_idents)]
+
+#[macro_use] extern crate common;
+#[allow(unused_imports)]
+#[macro_use] extern crate duct;
+#[macro_use] extern crate fomat_macros;
+#[macro_use] extern crate gstuff;
+#[macro_use] extern crate lazy_static;
+#[macro_use] extern crate serde_json;
+#[macro_use] extern crate serde_derive;
+#[macro_use] extern crate serialization_derive;
+#[macro_use] extern crate unwrap;
+
+#[path = "mm2.rs"]
+mod mm2;
+
+#[no_mangle]
+pub extern fn mm2_main() {
+    mm2::mm2_main()
+}
